feat: rename and add package info
